rename BERT to model
